Add proper room info response, and a join room request

+++ src/main/java/host/techcoop/gifthub/domain/UserVotes.java
@@ -6,5 +6,5 @@ import lombok.Value;
 public class UserVotes {
 
   String userName;
-  float percentSplit;
+  double percentSplit;
 }
